Struct amethyst_assets::AssetStorage

pub struct AssetStorage<A: Asset> { /* fields omitted */ }

An asset storage, storing the actual assets and allocating handles to them.

Methods

impl<A: Asset> AssetStorage<A>

pub fn new() -> Self

Creates a new asset storage.

pub fn unload_all(&mut self)

Remove all data from asset storages, invalidating all associated handles. Trying to retreive any data using old handle will return None.

pub fn clone_asset(&mut self, handle: &Handle<A>) -> Option<Handle<A>> where
A: Clone,

When cloning an asset handle, you'll get another handle, but pointing to the same asset. If you instead want to indeed create a new asset, you can use this method. Note however, that it needs a mutable borrow of self, so it can't be used in parallel.

pub fn get(&self, handle: &Handle<A>) -> Option<&A>

Get an asset from a given asset handle.

pub fn get_version(&self, handle: &Handle<A>) -> Option<u32>

Get an asset version from a given asset handle.

pub fn get_with_version(&self, handle: &Handle<A>) -> Option<&(A, u32)>

Get an asset and it's version from a given asset handle.

pub fn get_by_id(&self, id: u32) -> Option<&A>

Get an asset by it's handle id.

pub fn replace(&mut self, handle: &Handle<A>, asset: A) -> A

Replace asset under given handle, incrementing the version id. Returns old asset. Panics if asset handle is empty.

pub fn insert(&mut self, asset: A) -> Handle<A>

Insert preloaded asset into storage synchronously without going through usual loading step. You probably want to use Loader::load instead.

Use this method only when you need to insert procedurally generated asset directly into storage, skipping intermediate Asset::Data form.

pub fn contains(&self, handle: &Handle<A>) -> bool

Check if given handle points to a valid asset in the storage.

pub fn contains_id(&self, id: u32) -> bool

Check if given asset id points to a valid asset in the storage.

pub unsafe fn get_by_id_unchecked(&self, id: u32) -> &A

Get an asset by it's handle id without checking the internal bitset. Use contains_id to manually check it's status before access.

Safety

You must manually verify that given asset id is valid. Failing to do so may result in dereferencing uninitialized memory or out of bounds access.

pub fn get_mut(&mut self, handle: &Handle<A>) -> Option<&mut A>

Get an asset mutably from a given asset handle.

pub fn process<F>(
&mut self,
f: F,
frame_number: u64,
pool: &ThreadPool,
strategy: Option<&HotReloadStrategy>
) where
F: FnMut(A::Data) -> Result<ProcessingState<A>, Error>,

Process finished asset data and maintain the storage.

pub fn process_custom_drop<F, D>(
&mut self,
f: F,
drop_fn: D,
frame_number: u64,
pool: &ThreadPool,
strategy: Option<&HotReloadStrategy>
) where
D: FnMut(A),
F: FnMut(A::Data) -> Result<ProcessingState<A>, Error>,

Process finished asset data and maintain the storage. This calls the drop_fn closure for assets that were removed from the storage.