Struct amethyst_assets::Loader

pub struct Loader { /* fields omitted */ }

The asset loader, holding the sources and a reference to the ThreadPool.

Implementations

impl Loader

pub fn new<P>(directory: P, pool: Arc<ThreadPool>) -> Self where
P: Into<PathBuf>,

Creates a new asset loader, initializing the directory store with the given path.

pub fn with_default_source<S>(source: S, pool: Arc<ThreadPool>) -> Self where
S: Source,

Creates a new asset loader, using the provided source

pub fn add_source<I, S>(&mut self, id: I, source: S) where
I: Into<String>,
S: Source,

Add a source to the Loader, given an id and the source.

pub fn set_default_source<S>(&mut self, source: S) where
S: Source,

Set the default source of the Loader.

pub fn set_hot_reload(&mut self, value: bool)

If set to true, this Loader will ask formats to generate "reload instructions" which allow reloading. Calling set_hot_reload(true) does not actually enable hot reloading; this is controlled by the HotReloadStrategy resource.

pub fn load<A, F, N, P>(
&self,
name: N,
format: F,
progress: P,
storage: &AssetStorage<A>
) -> Handle<A> where
A: Asset,
F: Format<A::Data>,
N: Into<String>,
P: Progress,

Loads an asset with a given format from the default (directory) source. If you want to load from a custom source instead, use load_from.

See load_from for more information.

pub fn load_from<A, F, N, P, S: ?Sized>(
&self,
name: N,
format: F,
source: &S,
progress: P,
storage: &AssetStorage<A>
) -> Handle<A> where
A: Asset,
F: Format<A::Data>,
N: Into<String>,
P: Progress,
S: AsRef<str> + Eq + Hash,
String: Borrow<S>,

Loads an asset with a given id and format from a custom source. The actual work is done in a worker thread, thus this method immediately returns a handle.

Parameters

name: this is just an identifier for the asset, most likely a file name e.g. "meshes/a.obj"

format: A format struct which loads bytes from a source and produces Asset::Data with them

source: An identifier for a source which has previously been added using with_source

progress: A tracker which will be notified of assets which have been imported

storage: The asset storage which can be fetched from the ECS World using read_resource.

pub fn load_from_data<A, P>(
&self,
data: A::Data,
progress: P,
storage: &AssetStorage<A>
) -> Handle<A> where
A: Asset,
P: Progress,

Load an asset from data and return a handle.

pub fn load_from_data_async<A, P, F>(
&self,
data: F,
progress: P,
storage: &AssetStorage<A>
) -> Handle<A> where
A: Asset,
P: Progress,
F: FnOnce() -> A::Data + Send + Sync + 'static,

Asynchronously load an asset from data and return a handle.