Trait amethyst_assets::Progress

pub trait Progress {
    type Tracker: Tracker;
    fn add_assets(&mut self, num: usize);

    fn create_tracker(self) -> Self::Tracker;
}

The Progress trait, allowing to track which assets are imported already.

Associated Types

type Tracker: Tracker

The tracker this progress can create.

Loading content...
Required methods

fn add_assets(&mut self, num: usize)

Add num assets to the progress. This should be done whenever a new asset is put in the queue.

fn create_tracker(self) -> Self::Tracker

Creates a Tracker.

Loading content...