Struct amethyst_assets::Cache

pub struct Cache<T> { /* fields omitted */ }

A basic implementation for a cache. This might be useful as the Context of an Asset, so that the same asset doesn't get imported twice.

Because contexts have to be immutable, a RwLock is used. Therefore, all operations are blocking (but shouldn't block for a long time).

Methods

impl<T> Cache<T> where
T: Clone,

pub fn new() -> Self

Creates a new Cache and initializes it with the default values.

pub fn insert(&self, spec: AssetSpec, asset: T) -> Option<T>

Inserts an asset, locking the internal RwLock to get write access to the hash map.

Returns the previous value in case there was any.

pub fn get(&self, spec: &AssetSpec) -> Option<T>

Retrieves an asset, locking the internal RwLock to get read access to the hash map. In case this asset has been inserted previously, it will be cloned and returned. Otherwise, you'll receive None.

pub fn access<O, F: FnOnce(&T) -> O>(&self, spec: &AssetSpec, f: F) -> Option<O>

Accesses a cached asset, locking the internal RwLock to get read access to the hash map. In case the asset exists, f gets called with a reference to the cached asset and this method returns true.

pub fn retain<F>(&self, f: F) where
F: FnMut(&AssetSpec, &mut T) -> bool,

Deletes all cached values, except the ones f returned true for. May be used when you're about to clear unused assets (see Asset::clear).

Blocks the calling thread for getting write access to the hash map.

pub fn clear_all(&self)

Deletes all cached values after locking the RwLock.