Struct amethyst_assets::Directory

pub struct Directory { /* fields omitted */ }

Directory store.

Please note that there is a default directory storage inside the Loader, which is automatically used when you call load. In case you want another, second, directory for assets, you can instantiate one yourself, too. Please use Loader::load_from then.

Methods

impl Directory

pub fn new<P>(loc: P) -> Self where
P: Into<PathBuf>,

Creates a new directory storage.