Struct amethyst_assets::ProgressCounter

pub struct ProgressCounter { /* fields omitted */ }

A progress tracker which is passed to the Loader in order to check how many assets are loaded.

Methods

impl ProgressCounter

pub fn new() -> Self

Creates a new Progress struct.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn errors(&self) -> Vec<AssetErrorMeta>

Removes all errors and returns them.

pub fn num_assets(&self) -> usize

Returns the number of assets this struct is tracking.

pub fn num_failed(&self) -> usize

Returns the number of assets that have failed.

pub fn num_loading(&self) -> usize

Returns the number of assets that are still loading.

pub fn num_finished(&self) -> usize

Returns the number of assets this struct is tracking.

pub fn complete(&self) -> Completion

Returns Completion::Complete if all tracked assets are finished.

pub fn is_complete(&self) -> bool

Returns true if all assets have been imported without error.