Trait amethyst_assets::Format

pub trait Format {
    type Data;
    type Error: Error + Send + Sync;
    type Result: IntoFuture<Item = Self::Data, Error = Self::Error>;
    fn extension() -> &'static str;

    fn parse(&self, bytes: Vec<u8>, pool: &ThreadPool) -> Self::Result;
}

A format, providing a conversion from bytes to asset data, which is then in turn accepted by Asset::from_data. Examples for formats are Png, Obj and Wave.

Associated Types

type Data

The data type this format is able to load.

type Error: Error + Send + Sync

The error that may be returned from Format::parse.

type Result: IntoFuture<Item = Self::Data, Error = Self::Error>

The result of the parse method. Can be anything that implements IntoFuture.

Required Methods

fn extension() -> &'static str

Returns the extension (without .).

Examples

"png"

"obj"

"wav"

fn parse(&self, bytes: Vec<u8>, pool: &ThreadPool) -> Self::Result

Reads the given bytes and produces asset data.